+++ src/Classes/Files.php
@@ -21,6 +21,7 @@ use Contao\Input;
 use Contao\System;
 use Contao\CoreBundle\ContaoCoreBundle;
 use InvalidArgumentException;
+use Symfony\Component\Filesystem\Path;
 
 class Files
 {
@@ -250,5 +251,57 @@ class Files
     {
         $mime = strtolower($objFile->mime);
         return 'image/' === substr($mime, 0, 6) || 'application/pdf' === $mime;
+    }
+
+    /**
+     * Return the file extension from its name
+     * 
+     * @param string $name - file name
+     * 
+     * @return string - file extension
+     */
+    public static function getExtensionFromFilename(string $name): string
+    {
+        return substr($name, strrpos($name, '.') + 1);
+    }
+
+    /**
+     * Return the file absolute path
+     * 
+     * @param string $path
+     * 
+     * @return string
+     */
+    public static function getAbsolutePath(string $path): string
+    {
+        $strRootDir = System::getContainer()->getParameter('kernel.project_dir');
+
+        return Path::makeAbsolute($path, $strRootDir);
+    }
+
+    /**
+     * Return the file relative path
+     * 
+     * @param string $path
+     * 
+     * @return string
+     */
+    public static function getRelativePath(string $path): string
+    {
+        $strRootDir = System::getContainer()->getParameter('kernel.project_dir');
+
+        return Path::makeRelative($path, $strRootDir);
+    }
+
+    /**
+     * Return true if files exists
+     * 
+     * @param string $path
+     * 
+     * @return string
+     */
+    public static function exists(string $path): bool
+    {
+        return \file_exists(self::getAbsolutePath($path));
     }
 }